% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMorphologyData.R
\name{addMissingInfo}
\alias{addMissingInfo}
\title{Adds in values missing from each data type}
\usage{
addMissingInfo(storageList, animalIDs, treatmentIDs)
}
\arguments{
\item{storageList}{A list output by readInRawData}

\item{animalIDs}{A string vector of the animal IDs we want to read in data for}

\item{treatmentIDs}{A string vector of the treatment IDs we want to read in data for}
}
\value{
A list similar to storageList but with missing data added to the $Files elements
}
\description{
Adds in values missing from each data type
}
