% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bct_format.R
\name{bct.format}
\alias{bct.format}
\title{Produce formatted peak data table.}
\usage{
bct.format(data.in, long = T, QC.n = "QC", backup.obj = T, p.id = NULL,
  bn = NULL, seqm = T)
}
\arguments{
\item{data.in}{Input list object containing the peak data table to format as well as
an information table generated by bct.infogen.}

\item{long}{Logical; are the sample features in the rows (long = T). Transposes the
input data to contain features in columns.}

\item{QC.n}{Name of quality control sample group.}

\item{backup.obj}{Logical; should a backup object containing the remaining parts of
the input table after isolation of actual data be outputted?}

\item{p.id}{Column index of first sample column in input table (usually named 'Raw
abundance')}

\item{bn}{Names of batch labels within sample names (usually second row in input
table)}

\item{seqm}{Logical; should information on injection sequence extracted from sample
names be modified to become one continuing sequence? Needed when injection sequence
in the batches start at the same sequence index.}
}
\value{
Produces a list object with formatted peak table and backup object elements.
}
\description{
Function that takes the peak table input and produces the required format
  for passing in the bct correction algorithm. Also creates the required objects for
  reformatting to the input format.
}
