% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bct_msg.R
\name{bct.msg}
\alias{bct.msg}
\title{Console message printing}
\usage{
bct.msg(msg, side = c("both", "top", "bottom", "none"))
}
\arguments{
\item{msg}{Message text}

\item{side}{Broken line position. Can be "top" to print a line above the text, "bottom"
to print a line below the text, "both" to print lines above and below the text, or
"none" to not print lines.}
}
\value{
Prints a message to the R console.
}
\description{
Printing function for prettier messages. Not very fancy.
}
