% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell2loc_process.R
\name{cell2loc_process}
\alias{cell2loc_process}
\title{This function is to process the input data into the form required by cell2location}
\usage{
cell2loc_process(sc_data, st_data)
}
\arguments{
\item{sc_data}{this is a seurat object, assay = "RNA"}

\item{st_data}{this is a seurat object, assay = "Spatial"}
}
\value{
this function return the path of the processed data store,
 include "single_cell_metadata.csv", "single_cell_countdata.csv", "spatial_countdata.csv"
}
\description{
This function is to process the input data into the form required by cell2location
}
