% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produceIndex.R
\name{constructInfInd}
\alias{constructInfInd}
\title{Wrapper function. Takes the output of morphPreProcessing filtered for training conditions,
and returns a PCA object that can be used to build an Inflammation Index optimised for the TCS value and number of metrics
that lead to the greatest sensitivity to the training conditions}
\usage{
constructInfInd(
  procDat,
  method = "AUC",
  noDesc = 5:15,
  labCols = c("Animal", "Treatment", "TCSValue", "CellNo", "UniqueID"),
  correlationCutoff = 0.9
)
}
\arguments{
\item{procDat}{A data.table object output by morphPreProcessing that is filtered to only contains 2 Treatment values (training data)}

\item{method}{A string that indicates what method to use to select the optimal Inflammation Index. Can either be 'p value' or 'AUC'}

\item{noDesc}{An integer vector that indicates what different combinations of the best descriptors we want to try building our Index using}

\item{labCols}{A string vector of the ID columns in procDat}

\item{correlationCutoff}{A numeric value that indicates the correlation level at which we will remove metrics from our
inflammation index (to avoid building an index using overly correlated metrics)}
}
\value{
A list, where $PCA is a PCA object and $Metrics Correlation is a correlation matrix of the metrics the PCA
was run on, and $Optimal TCS is the optimal TCS value detected
}
\description{
Wrapper function. Takes the output of morphPreProcessing filtered for training conditions,
and returns a PCA object that can be used to build an Inflammation Index optimised for the TCS value and number of metrics
that lead to the greatest sensitivity to the training conditions
}
