% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMorphologyData.R
\name{convertHCtouM}
\alias{convertHCtouM}
\title{Converts FracLac metrics from pixels into the correct units to align with Parameter data}
\usage{
convertHCtouM(storageList, pixelSize)
}
\arguments{
\item{storageList}{A list output by retainCommonCells}

\item{pixelSize}{A numeric that is the pixel size in microns}
}
\value{
A list identical to storageList except with calibrated values for FracLac related data
}
\description{
Converts FracLac metrics from pixels into the correct units to align with Parameter data
}
