% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnDecon_main.R
\name{EnDecon_main}
\alias{EnDecon_main}
\title{This function is an ensemble model that integrates three methods of cell2location, RCTD, and spatialDWLS using a linear weighted model.}
\usage{
EnDecon_main(sc_data, st_data, python_path)
}
\arguments{
\item{sc_data}{this is a seurat object, assay = "RNA"}

\item{st_data}{this is a seurat object, assay = "Spatial"}

\item{python_path}{this is python path}
}
\value{
This function returns the result of four methods of deconvolution, including cell2location, RCTD, spatialDWLS, and EnDecon.
}
\description{
This function is an ensemble model that integrates three methods of cell2location, RCTD, and spatialDWLS using a linear weighted model.
}
