% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMorphologyData.R
\name{filterForVector}
\alias{filterForVector}
\title{Remove vector elements that don't match a filter, but retain original vector indices}
\usage{
filterForVector(filter_it, filter_by)
}
\arguments{
\item{filter_it}{A vector to be filtered}

\item{filter_by}{A string to find within the filter_it vector}
}
\value{
A vector where any elements that didn't contain filter_by are removed
}
\description{
Remove vector elements that don't match a filter, but retain original vector indices
}
