% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produceIndex.R
\name{filterTopMetrics}
\alias{filterTopMetrics}
\title{Filter our data to limit to the top performing parameters that have non-zero variance}
\usage{
filterTopMetrics(aggData, topParams, labCols)
}
\arguments{
\item{aggData}{A data.table output by formatROCRInput}

\item{topParams}{A string vector of metrics/parameters to retain}

\item{labCols}{A string vector of ID columns}
}
\value{
A data.table that is a filtered version of aggData
}
\description{
Filter our data to limit to the top performing parameters that have non-zero variance
}
