% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMorphologyData.R
\name{formatFracLacFiles}
\alias{formatFracLacFiles}
\title{Applies formatting to the FracLac data to align it with non FracLac data types}
\usage{
formatFracLacFiles(fracLacComboList)
}
\arguments{
\item{fracLacComboList}{A data.table object containing the FracLac data type}
}
\value{
A data.table object
}
\description{
Applies formatting to the FracLac data to align it with non FracLac data types
}
