% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produceIndex.R
\name{formatROCRInput}
\alias{formatROCRInput}
\title{Format morphPreProcessing Output for ROC-AUC Analysis}
\usage{
formatROCRInput(inputData, labCols)
}
\arguments{
\item{inputData}{A data.table output by morphPreProcessing}

\item{labCols}{A string vector of the ID columns (i.e. non-metric columns) in inputData}
}
\value{
A data.table that is a version of inputData where each metric has its own row
}
\description{
Format morphPreProcessing Output for ROC-AUC Analysis
}
