% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_sentiment.R
\name{get_tw_feelings}
\alias{get_tw_feelings}
\title{Positivity/negativity on Twitter.}
\usage{
get_tw_feelings(string, n_tweets = 20, since_date = Sys.Date() - 30)
}
\arguments{
\item{string}{The complete name of the university.}

\item{n_tweets}{The number of tweets to scrape.}

\item{since_date}{The date to first start scraping.}
}
\value{
A description of whether the university is viewed positively or negatively on Twitter. Scores vary from -4 to 4.
}
\description{
Positivity/negativity on Twitter.
}
\examples{
get_tw_feelings('Harvard University')
}
