% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_sentiment.R
\name{get_tw_words}
\alias{get_tw_words}
\title{Gets Twitter words.}
\usage{
get_tw_words(string, n_tweets, since_date)
}
\arguments{
\item{string}{The complete name of the university.}

\item{n_tweets}{The number of tweets to scrape.}

\item{since_date}{The date to first start scraping.}
}
\value{
A vector with the words from the tweets associated with the string.
}
\description{
Gets Twitter words.
}
