% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produceIndex.R
\name{getPC1PValue}
\alias{getPC1PValue}
\title{Calculate the p value of our Treatment effect on PC1}
\usage{
getPC1PValue(allDat)
}
\arguments{
\item{allDat}{A data.table object output by runPCA$allDat}
}
\value{
A list with two elements: model being the linear mixed model object we ran; pval being the p value of the effect of Treatment
in the model
}
\description{
Calculate the p value of our Treatment effect on PC1
}
