% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estsh.R
\name{gfpoly.multi}
\alias{gfpoly.multi}
\title{Polynomial mutiplication defined on GF(s) with a prime s}
\usage{
gfpoly.multi(f1, f2, s)
}
\arguments{
\item{f1}{a vector represents the coefficients of the first multiplier polynomial. For example, if the dividend is x^5+2x^3+1, then f1=c(1,0,2,0,0,1).}

\item{f2}{a vector represents the coefficients of the second multiplier polynomial. For example, if the divisor is x^4+2, then f2=c(1,0,0,0,2).}

\item{s}{a prime, the order of the Galois filed (GF).}
}
\description{
Polynomial mutiplication defined on GF(s) with a prime s
}
\examples{
gfpoly.multi(c(1,0,2,0,0,1),c(1,0,0,0,2),3)
}
