% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produceIndex.R
\name{hcCentreVariants}
\alias{hcCentreVariants}
\title{Return a string vector of our HC metrics and all their possible variants}
\usage{
hcCentreVariants(allMetrics)
}
\arguments{
\item{allMetrics}{A string vector of all possible metrics}
}
\value{
A string vector of all our HC metrics and their variants
}
\description{
Return a string vector of our HC metrics and all their possible variants
}
