% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produceIndex.R
\name{identifyMetricCorrelation}
\alias{identifyMetricCorrelation}
\title{For a input data table in long format containing 'Parameter' and 'Value' columns, return a data.table
that contains the number of other Parameters each Parameter correlates with above a given threshold value}
\usage{
identifyMetricCorrelation(inputDt, correlationCutoff)
}
\arguments{
\item{inputDt}{A data.table object containing Parameter and Value columns}

\item{correlationCutoff}{A numeric value that indicates the correlation level above which we want to identify metrics}
}
\value{
A data.table containing one row per unique Parameter in the input table and the number of other Parameters it correlates with
above our threshold value
}
\description{
For a input data table in long format containing 'Parameter' and 'Value' columns, return a data.table
that contains the number of other Parameters each Parameter correlates with above a given threshold value
}
