% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produceIndex.R
\name{identifyMetricVariants}
\alias{identifyMetricVariants}
\title{Takes a vector of metrics and returns a vector where each element contains the name of that metrics variant}
\usage{
identifyMetricVariants(allMetrics)
}
\arguments{
\item{allMetrics}{A string vector of metrics}
}
\value{
A string vector where the name of each element is a metric, and the value of each element is the name of the 
variant form of that metric
}
\description{
Takes a vector of metrics and returns a vector where each element contains the name of that metrics variant
}
