% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{InflammationIndex}
\alias{InflammationIndex}
\title{InflammationIndex: A package for analysing microglial morphology}
\description{
The InflammationIndex package contains functions used alongside the Microglia Morphology Analysis Fiji plugin
}
