% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_units.R
\name{inter_grid_voronoi}
\alias{inter_grid_voronoi}
\title{Computes intersections between grid and voronoi tesselation}
\usage{
inter_grid_voronoi(grid, voronoi, grid_id = "grid_id",
  voronoi_id = "NIDT", sf = FALSE, crs = 2154)
}
\arguments{
\item{grid}{data.frame of class \code{\link[sf]{sf}}.}

\item{voronoi}{data.frame of class \code{\link[sf]{sf}}.}

\item{sf}{logical; if \code{TRUE}, output is an sf object;
defauts to \code{FALSE}.}

\item{crs}{integer; desired projected coordinate system;
defauts to 2154 (Lambert 93).}
}
\value{
a simple data.frame if \code{sf=TRUE}. A data.frame of class
\code{\link[sf]{sf}} if \code{sf=FALSE}.
}
\description{
Computes intersections between a grid (e.g. created with
  \code{\link{create_grid}}) and a voronoi tesselation computed over an
  MNO antennas map. Intersections probabilities, defined as the relative
  area of an intersection in the correspond voronoi, are computed during
  the process.
}
\details{
The output data.frame contains a variable \code{proba_inter}
  which provides, for each intersection between a tile and a voronoi,
  the relative area of the tile in the corresponding voronoi. This
  value is comprised between 0 and 1 and thus assimilated to a
  probability of intersection.
}
\examples{
\dontrun{
# Create grid
grid_sf <- create_grid(x = fr_limits, tile_size = 500, export = FALSE)
# Import shapefile of voronoi tesselation
voronoi_sf <- sf::st_read("~/Antenne_voronoi_rev.shp", crs = 2154)
# Compute grid-voronoi intersections
table_prob <- inter_grid_voronoi(grid = grid_sf,
voronoi = voronoi_sf, prob = TRUE, sf = FALSE)
}
}
