% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estsh.R
\name{is.prime}
\alias{is.prime}
\title{Determine whether an integer is a prime}
\usage{
is.prime(x)
}
\arguments{
\item{x}{the integer to be determined.}
}
\value{
the result: TRUE (x is a prime) or FALSE (x is not a prime).
}
\description{
Determine whether an integer is a prime
}
\examples{
is.prime(7)
is.prime(8)
}
