% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobicount.R
\docType{package}
\name{mobicount}
\alias{mobicount}
\alias{mobicount-package}
\title{Improving present population counts using mobile phone data}
\description{
mobicount provides a flexible bayesian framework to improve present
  population counts using mobile phone data.
}
\author{
Romain Avouac
}
