% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMorphologyData.R
\name{morphPreProcessing}
\alias{morphPreProcessing}
\title{Wrapper function for all morphology data preprocessing for data output by the Microglia Morphology Analysis Fiji Plugin}
\usage{
morphPreProcessing(
  pixelSize,
  morphologyWD,
  TCSExclude = NULL,
  animalIDs,
  treatmentIDs,
  useFrac = F
)
}
\arguments{
\item{pixelSize}{A numeric identifying the pixel size of our analysed images in microns (assuming a square pixel)}

\item{morphologyWD}{A string path pointing to the 'Output' folder of the 'Working Directory' users used when running the Fiji plugin}

\item{TCSExclude}{OPTIONAL A vector of the TCS values (mask sizes) we don't want to read data in for. Defaults to NULL.}

\item{animalIDs}{A string vector of the animal IDs we want to read in data for}

\item{treatmentIDs}{A string vector of the treatment IDs we want to read in data for}

\item{useFrac}{OPTIONAL A boolean indicating whether we want to read in FracLac related data. Defaults to False.}
}
\value{
A data.table object of our morphology data merged and cleaned.
}
\description{
Wrapper function for all morphology data preprocessing for data output by the Microglia Morphology Analysis Fiji Plugin
}
