% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estsh.R
\name{onecoa.prime}
\alias{onecoa.prime}
\title{Generate a component orthogonal array (COA) with a prime d}
\usage{
onecoa.prime(d)
}
\arguments{
\item{d}{a prime, the column of the resulting COA.}
}
\value{
a COA with d(d-1) rows and d columns.
}
\description{
Generate a component orthogonal array (COA) with a prime d
}
\examples{
onecoa.prime(5)
}
