% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estsh.R
\name{onecoa}
\alias{onecoa}
\title{Generate a component orthogonal array (COA) with a prime power d}
\usage{
onecoa(d, p, f_d)
}
\arguments{
\item{d}{a power of prime p, the column of the resulting COA.}

\item{p}{a prime, the bottom number of d.}

\item{f_d}{a vector represents the coefficients of primative polynomial on GF(d). For example the primative polynomial on GF(3^2) is x^2+x+2, then let f_d=c(1,1,2).}
}
\value{
a COA with d(d-1) rows and d columns.
}
\description{
Generate a component orthogonal array (COA) with a prime power d
}
\examples{
onecoa(9,3,c(1,1,2))
}
