% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estsh.R
\name{onels}
\alias{onels}
\title{Generate an Latin square (LS)}
\usage{
onels(d)
}
\arguments{
\item{d}{an integer, the run size of the resulting LS.}
}
\value{
an LS with d rows and d columns.
}
\description{
Generate an Latin square (LS)
}
\examples{
onels(5)
}
