% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{pay_loans}
\alias{pay_loans}
\title{pay_loans}
\usage{
pay_loans(university, salary = 70000, interest = 0.03,
  per_sal = 0.15)
}
\arguments{
\item{university}{The complete name of the university.}

\item{salary}{The expected salary you will earn yearly in dollars.}

\item{interest}{The annual interest rate for the loans.}

\item{per_sal}{The percentage of the salary that would go into paying the loans.}
}
\value{
The number of years to pay entirely for the loans based on the tuition of the chosen university.
}
\description{
Time to pay student loans.
}
\examples{
pay_loans('Harvard University', salary = 80000, interest = 0.05, per_sal = 0.2)
pay_loans('Brown University', salary = 50000, interest = 0.01, per_sal = 0.1)
}
