% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{plot_city}
\alias{plot_city}
\title{plot_city}
\usage{
plot_city(university)
}
\arguments{
\item{schoolName}{The complete name of the college or university that is being plotted that must be present in the dataset.}
}
\value{
A map of the United States with a black dot of the college or university. A few summary statistics are given underneath the graph, such as whether the school is public or private, the school rank, location, and annual tuition.
}
\description{
Graphs a map of the United States and places a black dot where the inputted school is geographically located.
}
\examples{
plot_city(schoolName = 'Brown University')
plot_city(schoolName = 'Princeton University')
}
