% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_sentiment.R
\name{plot_moody_pie}
\alias{plot_moody_pie}
\title{Sentiment analysis on Twitter.}
\usage{
plot_moody_pie(string, n_tweets = 20, since_date = Sys.Date() - 30)
}
\arguments{
\item{string}{The complete name of the university.}

\item{n_tweets}{The number of tweets to scrape.}

\item{since_date}{The date to first start scraping.}
}
\value{
A pie chart plot with the sentiments associated with that univertity on Twitter.
}
\description{
Sentiment analysis on Twitter.
}
\examples{
plot_moody_pie('Princeton University')
}
