% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{plot_state}
\alias{plot_state}
\title{plot_state}
\usage{
plot_state(university)
}
\arguments{
\item{SchoolName}{The complete name of the college or university being plotted for the respective state that must be present in the dataset.}
}
\value{
A map of a single state in the US with an abbreviated label of the state pasted in the interior.
}
\description{
Graphs the an outline of the state that the inputted school is located in for the top 200 nationally ranked universities or colleges in the United States.
}
\examples{
plot_state(schoolName = 'Brown University')
plot_state(schoolName = 'Princeton University')
}
