% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors_computation.R
\name{prior_building}
\alias{prior_building}
\title{Computes prior from building registers}
\usage{
prior_building(paths_shp, grid, grid_id_var = "grid_id",
  dir_inter = paste0(tempdir(), "/dir_inter"), area_min_max = NULL,
  height_var = "HAUTEUR", impute = TRUE, height_min_max = c(NA, 130),
  parallel = TRUE, n_workers = 5, crs = 2154)
}
\arguments{
\item{paths_shp}{character vector; complete paths to
building shapefiles.}

\item{grid}{a \code{data.frame} of class \code{\link[sf]{sf}}.}

\item{grid_id_var}{character; name of the variable indicating tiles
unique identifier in \code{grid}.}

\item{dir_inter}{character; path to the desired output directory for
prior files.}

\item{area_min_max}{numeric vector of length 2; minimum and/or maximum
areas allowed for buildings, outside of which buildings are filtered.
To provide only one value, indicate the other as \code{NA}.}

\item{height_var}{character; name of the variable indicating building
heights. If set to \code{NULL}, prior is computed using only building areas.}

\item{impute}{logical; should heights of zero height buildings be imputed ?}

\item{height_min_max}{numeric vector of length 2; minimum and/or maximum
heights allowed for buildings, outside of which buildings are filtered.
To provide only one value, indicate the other as \code{NA}.
Minimum height is ignored if \code{impute = TRUE}.}

\item{parallel}{logical; if multiple building shapefiles are provided,
should the computations on each of these files be run in parallel ?}

\item{n_workers}{number of worker processes that \code{\link[sf]{doParallel}}
will use to execute tasks in parallel. Ignored if \code{parallel = FALSE}.}

\item{crs}{integer; desired projected coordinate system;
defauts to 2154 (Lambert 93).}
}
\value{
A \code{data.frame} of class \code{\link[data.table]{data.table}}
with two columns : tiles ID and a prior distribution over tiles.
}
\description{
Computes a prior distribution over a grid using data from building registers
(e.g. shapefiles of buildings over the territory).
}
\details{
If multiple shapefiles are provided (e.g. shapefiles at an
infra-national level), prior computation can be computed in parallel
using \code{\link[foreach]{foreach}}.
\cr \cr
If the provided grid was not computed using \code{\link{create_grid}},
tiles ID should be formatted the same way : "x_centroid/100:y_centroid/100".
}
\examples{
\dontrun{
# Import building shapefiles paths
bdtopo_shp <- list.files("~/bdtopo")
# Compute grid
grid_500_fr <- create_grid(x = fr_limits, tile_size = 500)
# Compute prior
build_prior_dt <- prior_building(paths_shp = bdtopo_shp, grid = grid_500_fr)
}
}
