% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produceIndex.R
\name{removeWorstPerformingVariants}
\alias{removeWorstPerformingVariants}
\title{Remove the worst performing variants from a string vector of metrics}
\usage{
removeWorstPerformingVariants(topParams, paramByAuc)
}
\arguments{
\item{topParams}{A string vector of metrics in paramByAuc}

\item{paramByAuc}{A data.table of the ROC-AUC value for all metrics}
}
\value{
A string vector that is a cleaned version of topParams
}
\description{
Remove the worst performing variants from a string vector of metrics
}
