% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMorphologyData.R
\name{retainCommonCells}
\alias{retainCommonCells}
\title{Remove all cells that aren't common across different data types}
\usage{
retainCommonCells(with_id)
}
\arguments{
\item{with_id}{A list output by formatUniqueID}
}
\value{
A list identical in structure to with_id but with non-common cells removed
}
\description{
Remove all cells that aren't common across different data types
}
