% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_sentiment.R
\name{setup_tweets}
\alias{setup_tweets}
\title{This function sets up the environment for twitter functions.}
\usage{
setup_tweets(consumer_key = "kmp7IUYAhZQftCrYkFSICCjuz",
  consumer_secret = "EMtBaYhcYV7V3vAEmSMSvApaSDJfj101fCD5TYAjicf4Z3ncy6",
  access_token = "1060241795240599557-Qyvx7TY0kYQ2ovPLYnMi6E4GNgirko",
  access_secret = "Q0AEBJvFx8kAh1p5tOCFppWppT12Mm8iEEj0zaf7dFvxL")
}
\arguments{
\item{consumer_key}{From API.}

\item{consumer_secret}{From API.}

\item{access_token}{From API.}

\item{access_secret}{From API.}
}
\description{
This function sets up the environment for twitter functions.
}
