% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_units.R
\name{sfc_as_cols}
\alias{sfc_as_cols}
\title{Transforms a sf POINT geometry to x,y columns}
\usage{
sfc_as_cols(x, names = c("x", "y"))
}
\arguments{
\item{x}{data.frame of class \code{\link[sf]{sf}} with a sfc_POINT geometry.}

\item{names}{character vector of length 2; names of the two new columns
with x,y coordinates; defauts to \code{c("x","y")}.}
}
\value{
A data.frame with two new columns corresponding to x,y coordinates
of spatial units. The initial geometry column is removed.
}
\description{
Transforms a data.frame of class \code{\link[sf]{sf}} with a POINT geometry
(e.g. a grid of which centroids has been computed using
\code{\link[sf]{st_centroid}}) to a conventional data.frame
 with coordinates as two numeric columns.
}
\details{
This transformation is useful for efficient filtering of coordinates,
  e.g. to subset a grid to a smaller area bounding box.
\cr
\cr
Source of the function : \url{https://github.com/r-spatial/sf/issues/231}
}
\examples{
\dontrun{
# Import grid and compute centroids
grid <- sf::st_read("~/grid_500_france.shp")
grid <- grid \%>\% st_centroid()
# Transform to a data.frame with centroid coordinates as new columns
grid_nogeo <- grid \%>\% sfc_as_cols()
}
}
