% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produceIndex.R
\name{shollSampledVariants}
\alias{shollSampledVariants}
\title{Takes a string vector of all the metrics we're analysing and returns a string vector of the
sholl percentile metrics that have (fit)}
\usage{
shollSampledVariants(allMetrics)
}
\arguments{
\item{allMetrics}{A string vector of metrics}
}
\value{
A string vector where each element is a sholl sampled metric that has a variant, and the name of the element
is its variant
}
\description{
Takes a string vector of all the metrics we're analysing and returns a string vector of the
sholl percentile metrics that have (fit)
}
