% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_sentiment.R
\name{show_tw}
\alias{show_tw}
\title{Show tweet.}
\usage{
show_tw(string, n_tweets = 1)
}
\arguments{
\item{string}{The name of the university to find a tweet about.}
}
\value{
A description of whether the university is viewed positively or negatively on Twitter. Scores vary from -4 to 4.
}
\description{
Show tweet.
}
\examples{
show_tw('Yale University')
}
