% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDWLS_pipeline.R
\name{spatialDWLS_pipeline}
\alias{spatialDWLS_pipeline}
\title{This function is to deconvolute ST data using the spatialDWLS method}
\usage{
spatialDWLS_pipeline(sc_obj, st_obj, python_path)
}
\arguments{
\item{sc_obj}{this is a seurat object, assay = "RNA"}

\item{st_obj}{this is a seurat object, assay = "Spatial"}

\item{python_path}{this is python path}
}
\value{
This function returns the result of spatialDWLS deconvolution
}
\description{
This function is to deconvolute ST data using the spatialDWLS method
}
