% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estsh.R
\name{structed.perm}
\alias{structed.perm}
\title{Generate the structured samples of simple random samples}
\usage{
structed.perm(permatrix, jcom, d)
}
\arguments{
\item{permatrix}{a matrix, each row is a permutation.}

\item{jcom}{an integer, represents the target component. Hope that the component jcom appears the same number of at each position.}

\item{d}{the number of components.}
}
\value{
a matrix represents the structured samples.
}
\description{
Generate the structured samples of simple random samples
}
\examples{
temp_samples<-matrix(nrow=10,ncol=5)
for(i in 1:10){temp_samples[i,]<-sample(1:5,5)}
structed.perm(temp_samples,3,5)
}
